--- a/src/PauliStrings.jl
+++ b/src/PauliStrings.jl
@@ -1,6 +1,6 @@
 module PauliStrings
 
-export Operator, OperatorTS1D, Operator64, Operator128, OperatorTS1D64, OperatorTS1D128, OperatorSorted
+export Operator, OperatorTS1D, Operator64, Operator128, OperatorTS1D64, OperatorTS1D128, OperatorSorted, OperatorUnsorted
 export trace, opnorm, eye, dagger, com, add, compress, ptrace, shift_left, shift, com, rotate
 export diag, xcount, ycount, zcount
 export truncate, trim, cutoff, prune, add_noise, k_local_part, participation