--- a/src/operations.jl
+++ b/src/operations.jl
@@ -202,7 +202,6 @@ function com(o1::Operator, o2::Operator; epsilon::Real=0, maxlength::Int=1000, a
     @assert typeof(o1) == typeof(o2) "Commuting operators of different types"
     o3 = Operator(o1.N)
     d = emptydict(o1)
-            k = (-1)^count_ones(o1.v[i] & o2.w[j]) - s * (-1)^count_ones(o1.w[i] & o2.v[j])
     @inbounds for i in eachindex(o1.v)
         v1, w1 = o1.v[i], o1.w[i]
         c1 = o1.coef[i]
@@ -210,6 +209,11 @@ function com(o1::Operator, o2::Operator; epsilon::Real=0, maxlength::Int=1000, a
             v2, w2 = o2.v[j], o2.w[j]
             v = v1 ⊻ v2
             w = w1 ⊻ w2
+            x1 = count_ones(v1 & w2)
+            k1 = 1 - ((x1 & 1) << 1)
+            x2 = count_ones(w1 & v2)
+            k2 = 1 - ((x2 & 1) << 1)
+            k = k1 - s * k2
             c = c1 * o2.coef[j] * k
             if (k != 0) && (abs(c) > epsilon) && pauli_weight(v, w) < maxlength
                 if isassigned(d, (v, w))