Hi Nicolas,

I had some spare time, and was trying to learn some bit-magic and experiment a bit with data structures. (In particular, reading through [this amazing document](https://en.algorithmica.org/), would recommend!)
To get a bit of practice, I figured might as well pick something useful to experiment on, so I played around a bit with the Pauli algebra.

What I really want to do is implement [this datastructure](https://en.algorithmica.org/hpc/data-structures/b-tree/), which I am not actually expecting to speed up this package at all.
However, to make my benchmarks somewhat fair, I first slightly optimized some of the algebra here. If you're interested in taking over some of that code, definitely more than welcome to cherry-pick, it needs a bit of love to actually fit it in the rest but I figured might as well already share.

I added some benchmarks, which you can run via https://github.com/MilesCranmer/AirspeedVelocity.jl:

```bash
benchpkg --rev={DEFAULT},c015ede87fe8,8c67ec8b9caa --output-dir=benchmark/results/ --bench-on=benchstart
```

|                            | main             | c015ede87fe8     | 8c67ec8b9caa     |
|:---------------------------|:----------------:|:----------------:|:----------------:|
| lanczos (N=50)/nterms=2^14 | 0.343 ± 0.0043 s | 0.263 ± 0.0048 s | 0.187 ± 0.038 s  |
| lanczos (N=50)/nterms=2^16 | 0.964 ± 0.0059 s | 0.701 ± 0.049 s  | 0.523 ± 0.019 s  |
| lanczos (N=50)/nterms=2^18 | 2.77 ± 0.017 s   | 1.89 ± 0.0056 s  | 1.4 ± 0.035 s    |
| lanczos (N=50)/nterms=2^20 | 6.07 s           | 3.79 ± 0.049 s   | 2.66 ± 0.019 s   |
| time_to_load               | 0.148 ± 0.0059 s | 0.147 ± 0.0065 s | 0.147 ± 0.0066 s |

---

In case you just want the low-effort optimizations: I checked with a profiler, and the main offender for this difference is this line:
```
k = (-1)^count_ones(o1.v[i] & o2.w[j]) - s * (-1)^count_ones(o1.w[i] & o2.v[j])
```
It seems like the compiler is not smart enough to realize that `(-1)^some_integer_power` simply needs to check the parity of the power, and it implements this by `power_by_squaring`, ie it really computes all the multiplications. This can be avoided by checking the parity with `(x & 1) in (0, 1)`, and then mapping it to desired `(-1, 1)` with `(1 - 2 * (x & 1))`, or in binary magic `(1 - (x & 1) << 2)`. For the commutator, ignoring that I renamed some of the variables this leads to the following:
```
k = ((count_ones(p2.v & p1.w) & 1) << 1) - ((count_ones(p1.v & p2.w) & 1) << 1)
```

---

I also toyed around for a bit with just keeping sorted vectors instead of the dictionaries, but it does seem like merging these is a lot more expensive (although this could presumably be multithreaded more easily Given my preliminary results, I didn't go into that further but the code is still here in case you wanted to play around with that yourself

---

Anyways, I did reorganize some of the code while I was trying things out, so the rest is a bit more of a significant change, but obviously free to adapt as you see fit or ignore ;)