@nicolasloizeau 2025-03-09
Hi, this looks awesome, thank you !
Could you please split this into two PR ? I'm happy to merge the benchmarks and low level optimizations that do not rely on the new OperatorUnsorted, OperatorSorted types. I need more time to understand these new data structures.