@lkdvos 2025-03-09
I'm sorry I might have not mentioned this clear enough: I don't really think this PR is good for merging at all, I just wanted to share some of the things I bumped into.
PS: I talked briefly to Dries (I'm also working at CCQ) and I understood that you'll be visiting soon-ish, so if you want we can discuss then. I'm not actually using PauliStrings.jl, so there's absolutely no rush on any of this from my part.