@nicolasloizeau 2025-03-09
ah ok, yes sure let's discuss in NY in a few weeks